[Profiler] Make KinetoEvent a view of Result (Part 1: trivial fields)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81319

Now that all data in the profiler flows through `torch::profiler::impl::Result` it is no longer necessary for `KinetoEvent` to maintain its own storage. Setting the fields on `KinetoEvent` has been a source of great consternation. It's very easy to forget to set a field for a particular case, and very hard to reason about what fields need to be set when.

We have to continue to support `KinetoEvent` because existing profiler machinery relies on it and (I think) it is part of the public API. That said, making it a simple view into a `Result` makes the semantics of each field much more clear.

This PR moves the fields which are simple views into one EventType (mostly TorchOp) and uses a default value elsewhere. Later PRs will tackle the thornier aspects.
ghstack-source-id: 163655370

Differential Revision: [D37481559](https://our.internmc.facebook.com/intern/diff/D37481559/)